+++ packages/react/src/icons/Voicemail.tsx
@@ -1,3 +1,3 @@
 import * as React from "react";
 import type { SVGProps } from "react";
-export const Voicemail = (props: SVGProps<SVGSVGElement>) => <svg width="1em" height="1em" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}><rect x={3} y={9} width={6} height={6} rx={3} stroke="black" strokeWidth={2} /><rect x={15} y={9} width={6} height={6} rx={3} stroke="black" strokeWidth={2} /><rect x={19} y={14} width={2} height={14} rx={1} transform="rotate(90 19 14)" fill="black" /></svg>;
+export const Voicemail = (props: SVGProps<SVGSVGElement>) => <svg width="1em" height="1em" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}><rect x={3} y={9} width={6} height={6} rx={3} stroke="currentColor" strokeWidth={2} /><rect x={15} y={9} width={6} height={6} rx={3} stroke="currentColor" strokeWidth={2} /><rect x={19} y={14} width={2} height={14} rx={1} transform="rotate(90 19 14)" fill="currentColor" /></svg>;